ADD: canImport and canExport in schema tests - CHANGE: import-export.js file directory (moved to common folder)
