+++ lib/schemas/common/browseBase/browseBase.js
@@ -9,7 +9,7 @@ const getEndpointParameters = require('../endpointParameters');
 const filters = require('../../browse/modules/filters');
 const sortableFields = require('../../common/sortableFields');
 const statusBar = require('../../common/statusBar');
-const makeExportImportSchema = require('./import-export');
+const makeExportImportSchema = require('../import-export');
 
 const PAGE_SIZES = [15, 30, 60, 100, 'none'];
 const DEFAULT_PAGE_SIZE = 60;
