+++ tests/mocks/schemas/planning.yml
@@ -3,6 +3,8 @@ name: planning
 root: Planning
 title: planning
 canRefresh: false
+canImport: true
+canExport: true
 filters:
   - name: filterInput
     label: someLabel
